+++ injector/network_latency.go
@@ -12,7 +12,9 @@ import (
 
 	"github.com/DataDog/chaos-controller/api/v1beta1"
 	"github.com/DataDog/chaos-controller/container"
+	"github.com/DataDog/chaos-controller/metrics"
 	"github.com/DataDog/chaos-controller/network"
+	"github.com/DataDog/chaos-controller/types"
 	"go.uber.org/zap"
 )
 
@@ -32,13 +34,13 @@ type NetworkLatencyInjectorConfig struct {
 }
 
 // NewNetworkLatencyInjector creates a NetworkLatencyInjector object with the default drivers
-func NewNetworkLatencyInjector(uid string, spec v1beta1.NetworkLatencySpec, ctn container.Container, log *zap.SugaredLogger) Injector {
-	return NewNetworkLatencyInjectorWithConfig(uid, spec, ctn, log, NetworkLatencyInjectorConfig{})
+func NewNetworkLatencyInjector(uid string, spec v1beta1.NetworkLatencySpec, ctn container.Container, log *zap.SugaredLogger, ms metrics.Sink) Injector {
+	return NewNetworkLatencyInjectorWithConfig(uid, spec, ctn, log, ms, NetworkLatencyInjectorConfig{})
 }
 
 // NewNetworkLatencyInjectorWithConfig creates a NetworkLatencyInjector object with the given config,
 // missing fields being initialized with the defaults
-func NewNetworkLatencyInjectorWithConfig(uid string, spec v1beta1.NetworkLatencySpec, ctn container.Container, log *zap.SugaredLogger, config NetworkLatencyInjectorConfig) Injector {
+func NewNetworkLatencyInjectorWithConfig(uid string, spec v1beta1.NetworkLatencySpec, ctn container.Container, log *zap.SugaredLogger, ms metrics.Sink, config NetworkLatencyInjectorConfig) Injector {
 	// traffic controller
 	if config.TrafficController == nil {
 		config.TrafficController = network.NewTrafficController(log)
@@ -57,8 +59,10 @@ func NewNetworkLatencyInjectorWithConfig(uid string, spec v1beta1.NetworkLatency
 	return networkLatencyInjector{
 		containerInjector: containerInjector{
 			injector: injector{
-				uid: uid,
-				log: log,
+				uid:  uid,
+				log:  log,
+				ms:   ms,
+				kind: types.DisruptionKindNetworkLatency,
 			},
 			container: ctn,
 		},
@@ -152,6 +156,7 @@ func (i networkLatencyInjector) Inject() {
 		// retrieve link from name
 		link, err := i.config.NetlinkAdapter.LinkByName(linkName)
 		if err != nil {
+			i.ms.MetricInjected(i.container.ID(), i.uid, false, i.kind, []string{"link:" + linkName})
 			i.log.Fatalf("can't retrieve link %s: %w", linkName, err)
 		}
 
@@ -167,6 +172,7 @@ func (i networkLatencyInjector) Inject() {
 				clearTxQlen = true
 
 				if err := link.SetTxQLen(1000); err != nil {
+					i.ms.MetricInjected(i.container.ID(), i.uid, false, i.kind, []string{"link:" + link.Name()})
 					i.log.Fatalf("can't set tx queue length on interface %s: %w", link.Name(), err)
 				}
 			}
@@ -179,12 +185,14 @@ func (i networkLatencyInjector) Inject() {
 			priomap := [16]uint32{1, 2, 2, 2, 1, 2, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1}
 
 			if err := i.config.TrafficController.AddPrio(link.Name(), "root", 1, 4, priomap); err != nil {
+				i.ms.MetricInjected(i.container.ID(), i.uid, false, i.kind, []string{"link:" + link.Name()})
 				i.log.Fatalf("can't create a new qdisc for interface %s: %w", link.Name(), err)
 			}
 		}
 
 		// add delay
 		if err := i.config.TrafficController.AddDelay(link.Name(), parent, 0, delay); err != nil {
+			i.ms.MetricInjected(i.container.ID(), i.uid, false, i.kind, []string{"link:" + link.Name()})
 			i.log.Fatalf("can't add delay to the newly created qdisc for interface %s: %w", link.Name(), err)
 		}
 
@@ -193,6 +201,7 @@ func (i networkLatencyInjector) Inject() {
 		if len(ips) > 0 {
 			for _, ip := range ips {
 				if err := i.config.TrafficController.AddFilterDestIP(link.Name(), "1:0", 0, ip, "1:4"); err != nil {
+					i.ms.MetricInjected(i.container.ID(), i.uid, false, i.kind, []string{"link:" + link.Name()})
 					i.log.Fatalf("can't add a filter to interface %s: %w", link.Name(), err)
 				}
 			}
@@ -203,10 +212,21 @@ func (i networkLatencyInjector) Inject() {
 			i.log.Infof("clearing tx qlen for interface %s", link.Name())
 
 			if err := link.SetTxQLen(0); err != nil {
+				i.ms.MetricInjected(i.container.ID(), i.uid, false, i.kind, []string{"link:" + link.Name()})
 				i.log.Fatalf("can't clear %s link transmission queue length: %w", link.Name(), err)
 			}
 		}
 	}
+
+	i.ms.EventWithTags(
+		"network latency injected",
+		"network latency injected",
+		[]string{
+			"ContainerID:" + i.container.ID(),
+			"UID:" + i.uid,
+		},
+	)
+	i.ms.MetricInjected(i.container.ID(), i.uid, true, i.kind, []string{})
 }
 
 // Clean cleans the injected latency
@@ -245,4 +265,13 @@ func (i networkLatencyInjector) Clean() {
 			i.log.Fatalf("can't delete the %s link qdisc: %w", link.Name(), err)
 		}
 	}
+
+	i.ms.EventWithTags(
+		"network latency cleaned",
+		"network latency cleaned",
+		[]string{
+			"ContainerID:" + i.container.ID(),
+			"UID:" + i.uid,
+		},
+	)
 }